Initial Renderer
For now this repo is going to be a nice little sandbox to stomp around in.
I guess this repo's commit history is just going to be a diary.

The Renderer crate holds a high level engine that safely contains ash,
the vulkan rust bindings. With this commit the core renderer can be initialized,
the next step is going to be recording commands and actually drawing things. The
code is extremely close to the ash examples, with some code just straight up being
borrowed. The ash examples are great and are a solid launching platform.

My plan is to get vulkan off the ground, so I can get some real graphical
programming experience before going to nvidia. I plan on fooling around with some
complicated renderings that have nothing to do with desktop environments. This
will be a good test for me to learn benchmarking and optimizations in the wild.

Next/concurrently I will be messing with wayland to get a feel for it. At some
point when that gets merged in there are going to be some interesting repo
restructurings.

Finally, this code lives on sourcehut because I want to see if it is
a platform I enjoy and want to commit to in the future. So far so good.

One day this repo will get nuked and a real codebase will be constructed from
the fallout.