Add the simplest possible wayland compositor
Reorganization:
* vulkan - vulkan rendering practice
* wayland-server - a very basic wayland compositor

This new compositor was created to test what life is like
using raw wayland bindings, and experimenting with unsafe
rust. Wayland is a very asynchronous, callback friendly library
which does not pair smoothly with unsafe rust.

The compositor uses the `minifb` crate to create an X11 window,
in which surfaces can be rendered in an arbitrary Z-order. The
order of windows depends on which one is added first in the
list. Windows are also only made at the top left of the screen
and can never be moved. It's not meant to look nice, just to
experiment with different programming models in wayland.

Most of the wayland resources are *mut types which is not
safe at all. My first attempt at making this safer is to
ahve all of the private structures handed back by wl_resources
be a RefCell<>. The programming paradigm looks like this:

```
// from Compositor::create_surface

// get the refcell that we set as the private data
let cell: *mut RefCell<Surface> = unsafe {
    wl_resource_get_user_data(resource) as *mut RefCell<Surface>
};

// Important: check that the pointer is not null
if cell == ptr::null_mut() {
    return;
}

// Now we can borrow a mutable reference. This can panic
let mut surface = unsafe { (*cell).borrow_mut() };
```

The idea is to use a refcell (or a better multithread friendly
primitive) to arbitrate access to a `Compositor` or `Surface`
struct. This _should'nt_ panic in single threaded operation,
but it is a nice safeguard.

The next order of business (on the wayland side) is to get a nice
profiling setup going. I want to see what the performance, and security,
of the code I've been writing is like. I also want to profile some
resource heavy wayland apps on linux to get a feel for where most of
the computation time is spent. I hope to do this on elementary OS, as
it looks quite nice.

Other tasks would include implementing one of the simpler shell
protocols to help with getting actual windows done in the compositor.
This should be very easy.