Finish up the vulkan tutorial
A rainbow triangle can now be displayed on mac.

The render loop looks like this:

Renderer.start_frame  --> Draw our scene
Renderer.present      --> Present the frame to the swapchain

The idea is to separate the frame drawing and presentation, so
that any post-processing can be done inbetween.

With this commit the basic renderer is done, and I can start
playing around with complicated scenes and shader types. One of
the big omissions from this commit is vertex and index buffers.
Right now the geometry is just created in the vertex shader.

The next things on the list are:
* Vertex and index buffers
* Try out renderdoc
* Create perspective transformation
* Figure out the optimal memory type strategy

One of my concerns is that ash has non-zero overhead that will
be a concern in a battery environment. I need to make sure that
the release build has optimizations that inline the ash functions
or something. Obviously I need to profile this first to make sure
it is actually a problem.