Support drawing arbitrary obj files
The main 3D display pipeline is all set up. It is fairly limited
at the moment, but it's a solid base to build on. This is a good
point to fork for scraping the 3D portions when I go back to writing
the compositor.

The obj-rs crate is used to load objects from blender. Blender models
should be exported with a Y-up and a Z-forward in the wavefront form.

The Renderer keeps a list of meshes which have been added like so:
```
rend.add_mesh(
	obj_vertices.as_slice(),
	obj_indices.as_slice(),
);
```

Each object to render is held in the `Mesh` struct, which holds
resources for the vertex and index buffers, along with materials
and other extras which will be added later.

Uniform buffers are used (one for each swapchain) for passing the
model/view/projection matrices which will be applied in the vertex
shader.

All objects must be in indexed vertex format. obj-rs also requires
that they have vertex normals or else it will panic.

Things needed:
* Push constants for shader materials
* Initial offsets for meshes
* Descriptor updates for object transformations
* Camera movement