Added diffuse and ambient shading
Diffuse shading is calculated based on the dot product
of the face's normal vector and a direction to the light source.
Both of these need to be normalized or else the color will be off.

The light position should be implemented as a push constant. This
will be part of the next big sweeping lighting changes. Specularity
will also be added into the shader.

An interesting thing I learned is that data pertaining to the vertices
is passed "in band" with the vertex data itself. The vertex input attachment
descriptions specify the different components of the per-vertex data.
The current setup for binding 0 looks like this:

Attachment 1 - vertex position vector
Attachment 2 - normal vector

The attachments can then be read from the shader as:
layout(location = 0) in vec3 loc;
layout(location = 1) in vec3 norm;