Support for x11 on FreeBSD and fix memory types
Add in the create_surface method for xlib surfaces.
This has an unusual quirk on freebsd where the standard
validation layers do not support the x11 surfaces. Validation
layers are disabled on the unix platform(s) as a result.

Also fix the find_memory_type_index function. The memory requirements
actually specify a bitmask for entries in the memory properties' list
of memory types:

According to the spec:
memoryTypeBits is a bitmask and contains one bit set for every
supported memory type for the resource. Bit i is set if and only
if the memory type i in the VkPhysicalDeviceMemoryProperties
structure for the physical device is supported for the resource.