Rotate the scene and update the lighting/models
Create a new helper to update the contents of an area of memory.
The memory is updated with a new rotation matrix, the descriptor
set does not need to be rebound.

An interesting thing I learned is that you do not have to borrow
self as mutable to call methods on the device. The device level
methods all borrow self as immutable, so I guess concurrent
operations are allowed.

The light position was also changed, and the model was updated.
Unfortunately I could not get smooth shading to work exporting
from blender.