Add a basic demo of libinput (and udev)
libinput sits on top of udev, which sits on top of devd. Together
they allow for the processing of evdev input sources, usually from
`/dev/input/event*`

BSD does not have systemd, so udev is provided by libudev-devd, which
is just a wrapper for the normal devd.

This program uses the input and udev crates to provide basic reading
of events. This is about as basic of an example as possible, all it
does is print the available input devices and prints any events it
receives.