DisplayKHR works on physical screen
- Teacup spins happily on the old toshiba -

This is version one of the raw display code. It's version
one because I don't do any selection based on the properties
of the display, I just grab the first one. Somehow this works
out in the end.

There aren't a lot of comments in this commit because I want
to redo the way that the display is generated. I need to read
that portion of the spec again and my comments will be much clearer
after doing all of that.

This took a small hiatus into fixing the i915kms driver:
* Needed to have mesa's anv use the unsynchronized gem userptr.
* The worker task for i915kms's __i915_gem_userptr_get_pages_worker
  needed a section un-ifdef'd out.

I also updated the suite of vulkan related ports to 1.1.130. None
of this stuff has been contributed back yet.

Next on the list is fixing the display properties and resolution.