Add a cbuf for each frame
The cbufs array in Renderer will now be as long
as the swapchain image count. Each image has its
drawing operations recorded beforehand in its own
framebuffer. This should ideally be no more than
three images.

Each frame the shader constants are updated with
a new rotation amount and the existing command
buffers are dispatched. I did the cardinal sin
of not benchmarking this before and after, so I
don't have a good measure of its effect.