Add a macro to get userdata from a wl_resource
See the extensive comment above `get_userdata`. Other fixes
include:
* Removed minifb (this breaks rendering)

The Compositor does not use the normal RefCell approach when it
is passed as userdata. The Compositor object is used to dispatch
all of the event calls, and so it would always be borrowed as mut
and fail the dynamic checking. It is just grabbed from its raw
pointer.

This could pose problems down the line if I am doing multithreaded
dispatch. Need to look into this.