Texture images are almost there
At the moment this does not work. Many things
have been totally reworked, including image
transitions for images with texture contents.

The cbuf recording is still split from the render
loop, so some investigation is needed to see if that
is causing any issues.

Currently one quad is displayed in the bottom right, but
it does not have a texture. I seemed to have followed
the tutorial correctly but something has gotten mixed up.