Get images displayed on screen
Finally the major rewrite is here. Lesson learned: ALWAYS
go to the validation layers first.

Important TODO:
* Get validation layers working on FreeBSD
* Only allocate one image sampler per framebuffer
  * One sampler per frame, record rebinding the descriptors for each window

There are too many changes to write about here. Mostly I just fixed
a lot of errors which had been around for a while. The samplers are
based on the vulkan tutorial, but are quite different at this point
in time.

Also an important thing I learned is that descriptor sets bound separately
need to be referenced separately by the `set` qualifier in glsl:

layout(set = 1, binding = 1) uniform sampler2D tex;

This is because the image sampler is bound as the second descriptor
set. This seems to be a common source of confusion in vulkan