Add push constants for depth
Each window has 2D vectors of its location, but
a z-ordering is needed to lay out windows in a
front to back order. Push constants take care of
this.

`record_draw` will need to push a new ordering
value before drawing the window. This should
be relatively cheap.

This sorting is paving the way for a desktop
background and multiple windows.