Rework meshes and composite with titlebars
Technically we now have a horrible compositor!

This reworks the mesh interface quite a bit, adding
a method for recording draw calls for any mesh, and
adding a new concept called an App (which should
really be renamed to window and not have a vec<mesh>)

A context holds multiple apps, which are just onscreen
client windows. Anything drawn on screen is represented
by a mesh, including the desktop background.

A big overhaul needs to happen with cbuf recording,
tesselation shaders, and image samplers. Before I
can do that I need to find a good way to profile
this stuff on bsd.

Note: depth needs to be in [0, 1.0)