Fix destructors and update for profiling
This fixes the Display destructor to actually be called.

Also records the cbuf for the next frame so that it is
recorded in the dtrace profiling. This is worse for
performance, but I want to see what it's impact is.

Also there is a new command line argument for timed
execution.

Usage: vulkan [timed]

Using this will print out the runtime and FPS for
some number (8000) iterations. Useful to make sure
the performance isn't slipping.