Add a descriptor pool allocator
One of the things I am most displeased with is that
each window requires a sampler descriptor for every
swapchain image. This adds up quick.

This commit adds DescPool, a pool of descriptor pools.
The mesh creation simply asks this allocator for the
sets, and DescPool keeps track of allocating new pools
as needed.

Right now each pool holds only 4 sets, which is way
too little. This is so that I can quickly trigger
the creation of many pools.