Bring vulkan and wayland under the same roof
Finally it is starting to take shape. Compositor
is a new directory holding a more appropriate
organizational structure. vkcomp is the vulkan
compositor which holds the Renderer, and ways
is the wayland server.

Next step is to have ways use vkcomp for drawing
the windows it gets. This should be wrapped into
a compositor module.