Wayland shm clients can now be composited
For the first time, this is actually a wayland compositor.

This is the result of a large refactor adding in the Task
API for inter-subsystem communication. As of now there
are only three subsystems.

Category5:
This is the top level subsystem, and it is only in
charge of starting the other ones. It will eventually
do fancy management things like changing settings and
restarting subsystems.

ways:
This is the wayland server subsystem. It tells the
renderer when new windows and surfaces are created
or destroyed.

wm:
This is the window manager, also known as the renderer.
It listens for updates of windows (Apps) based on their
id, and handles anything graphical.

Subsystems speak to each other over multi-producer
single-consumer channels. This provides some clear
ownership boundries, and means that there is a clear
division between subsystems. All data passed between
subsystems uses the 'Task' API. A Task is a structure
outlining a body of work that the subsystem needs
to do. The only real one so far is in wm/task.rs