Windows can now be deleted
Support the wl_surface.delete request now.

Some wonky stuff had to be done for shm Meshes.
Because they are used for all of the frames,
they need to not be recorded in the next frame
before they can be destroyed.

Add a reaper system, where Apps are marked for
death, not recorded, and then reaped during the
next available downtime.

The structure of the worker thread will probably
need to change in the future so we aren't only
doing one event per frame.