Add wl_shell skeleton and redo ways
Added a skeleton for the wl_shell interfaces,
and updated the way macros and ways dispatching
works

There is a new struct, EventManager, which is
in charge of holding the singleton and event
loop stuff. It is a box that holds the compositor
interface.

This is a much better split. Also update the way
that bindings are built to add an include path.