WIP of a big rewrite to wayland_server
My attempts to roll by own wayland bindings was not going
well, it was very unsafe and hard to follow. The final nail
in the coffin was the lack of support for protocol extensions,
which I should have realized sooner.

Moving to the wayland_server crate because it provides two
out of those three things. Safety and protocol extensions are
big power-ups to the wayland stack, although I think their
API is dreadful to write in.

At the moment I've gutted most of the wayland code, and this
(untested) commit rewrites the compositor interface in
wayland_server