Add a TimingManager
The motivation for this is frame callbacks, which
need to take place once every 16 ms (once a frame
at 60 fps). This struct keeps track of how much
time is remaining before an action needs to be called,
and callers can use this number for their timeout
values.

This isn't a timing subsystem, but rather a helper
for tracking timing information.