Add logging infra and a stopwatch
For the forseeable future I will be flying blind performance-wise
since dtrace's ctfconvert doesn't look like it can process rust.
The stopwatch should fill the gaps, it'll be useful for printing
how long portions of hot loops take.

The logger has different categories, the main downside is importing
it is really annoying. A todo is to add a prelude to make that easier.

These two helped expose the current issue: nvidia seems to block
for the next image in a frame in a different place than mesa. I
think it blocks on the presentation sema instead of image acquisition.