Fix hemisphere replaying
Patches need to be replayed onto hemispheres to keep them up to
date. It's just like damage tracking between two double buffers.

This also sorts out some synchronization and logging problems. The
end result is the same functionality as pre-atmosphere but with
an entire order of magnitude less cpu usage.