Fix task queueing and fd duplications
This fixes issues with vkcube

We need to check the result of duplicating the dmabuf fd and do nothing
if the fd is bad. If the client program was killed this descriptor is no
longer valid so just ignore it.

Also do not use a Vec for queueing tasks, we need a VecDeque (which is
a proper queue).