WIP: add window properties to atmosphere
At this point all the functionality is back. There are some issues,
primarly with how the trees are stored in RwLocks which are not
consistent with the hemisphere state. These need to be changed to
be properties and this should go away.

There is also the problem of needing to initialize properties so
that their getters don't panic on None values. This will be annoying
going forward