Make skiplist properties
This replaces the RwLock approach that I experimented with. A locked
tree was removed because it faced efficiency and consistency issues,
mainly being that it was too hard to keep in sync with the props.

Now the skiplist is implemented as a set of properties in skiplist.rs,
which is much more convenient. Writing the linked list resulted in
some weird corner cases, but overall the experience was pretty good
and I consider the property system a success.