Implement wl_subcompositor
This is mostly untested because I haven't done that sooo

This fully implements the subcompositor and subsurface interfaces, which is the
next step towards getting winit apps up and running. This reuses the skiplist
code to make a list of subsurfaces for each surface.

Rendering recursive subsurfaces is currently a problem, that needs to be changed
in wm. Also need to experiment with sync and desync.

Either way it's a good first step. The next thing to look at is resizing windows
which is going to be brutal. After that I'll look at OpenGL apps and at that
point hopefully a real application will be runnable