Correctly track window size based on ack
This properly sets the window size during surface commit to whatever the last
acked configuration event was. This is tracked in the TLConfig struct, which
holds both the size at the time of this config event and the state.

The size from the TLConfig struct may be overwritten by the double-buffered
state from xdg_surface.set_window_geometry. This seems like a problem to me, but
I think that's what the spec intends. It doesn't seem to happen irl.

The current rendering problems are due to vkcomp operating based on the *window*
sizes instead of the *surface* sizes. See my notebook for 11/10/20 for the plan