Fix dmabuf clients
This adds additional supported format querying to the dmabuf import code. It
also creates the image with the correct size and modifier, instead of just using
a one-size-fits-all.

I also switched the color format to vk::Format::B8G8R8A8_SRGB, which fixes the
incorrect colors being displayed for vkcube. I probably need to re-evaluate the
modifiers we advertise and actually check that the passed dmabuf params are
supported.