Finally complete smooth resizing
weston-terminal and weston-resizor are finally working!

The problem was that the image needed to be destroyed and re-created every time
it was resized. For some reason when a large buffer was copied into an image
that was too small, instead of having a problem it just silently only copied not
enough data into the image. There is probably some room for optimization here.

There is also some room for memory safety, since I noticed that there is a
memory leak from when clients are closed.