Big reorganization and creation of thundr
This splits the one gigantic crate into a few logical ones.

category5 - The desktop environment
	vkcomp - the window manager
	ways - the wayland server
	input - the libinput handler
	atmosphere - the entity-component set connecting vkcomp and ways
thundr - the renderer
utils - common structs and handlers for bot

Tested the apps, no regressions except for vkcube being translucent (because I
changed the draw order?). Continuing with the thundr rework will modify this so
I'll ignore it for now