Move AppContext -> GeomPipeline
At this point there is a complete isolation of the pipeline that renders the
windows from the base renderer.

i.e. Renderer is the common bits of the renderer, and a Pipeline will use it to
draw frames. The geometric (GeomPipeline) is the traditional renderer that
paints windows on quads in 3D space.