Fix compute cpu perf and descriptor indexing
- move away from hashmap for keeping track of in-use tiles. Hashmap
was performing terribly and my custom solution is significantly better.
With this the cpu % is around 6-12% while moving the mouse, down from 50+

- Use the correct flags and pdev features for bindless
PARTIALLY_BOUND was the important missing piece. See:
https://gist.github.com/DethRaid/0171f3cfcce51950ee4ef96c64f59617

With the validation layers on I also had to revert the swizzle change
so the b & r channel problem still persists