Notify wm of created windows immediately
With the old scheme it was possible for windows to be created, surface
contents updated, and xdg set toplevel to occur. The surface contents would
get lost because wm wasn't nofied of the window untile toplevel was set.

Now send the create_window task immediately, so wm can track surface updates