Make SurfaceList an rc wrapped type
We need this so that the surfaces can reference the main list and
add any damage they have to it. This includes many gross violations
of the law of demeter due to the fact that I can't return the iterator
like I could when there wasn't a refcell in play.