Split out PropertyList
A PropertyList is a vector that keeps one single property for each
PropertyId. PropertyMap is implemented on top of this by having that
one property be a vector of n properties.

This also removes the app list in vkcomp, which should help performance
since we don't have the scan the vec to do anything.