wayc: Add surface/role/buffer skeletons and start using anyhow
Anyhow seems to be the nicest way to propogate errors, and for once I don't
feel like rolling my own. Anyhow is reexported in the root of wayc, and it
should be used from there to allow for overriding in the future.

This adds skeletons for surface/role/buffer, and starts hammering out the
initial api.