dakota: add layout tree
This adds an IR called LayoutNode, which will trim an Element into
the essentials before we create a thundr surface. This offers a
private translation layer we can optimize before thundr.

Might have to remove in the future, depending on if this is more/less
efficient than using thundr surfaces which are Rc