thundr: Use transfer buffer for loading shm images
This fixes a correctness issue in nv, where there was a slash through
weston-terminal. This fixes it by appropriately using the correct
layout, and actually copying images in a valid manner.

Also cleans up some code to make things readable.

Tested with cat5 on nv/intel, thundr-test on nv, dakota-test on intel
breaks scene1 on nv