thundr: Don't initialize the transfer mem
Very weird on nvidia: The free_memory call in upload_memimage_to_transfer
causes nvidia presentation to get messed up in a diagonal. Not reproducible
in renderdoc. Not initializing the transfer mem is a WAR, but I don't
understand why this would happen.

Removing the call for initialize_transfer_mem fixes everything