dakota: Guard Element with Rc<RefCell<<>>
This is unfortunately the best way for us to hand out references
to the elements in a tree. We need this to safely keep elements in
a) the dom, b) each attached event handler, and c) app references
to individual elements.

This is done with serde's rc feature, since serializing/deserializing
will generate an element each time it sees an Rc, it doesn't deduplicate.
This is fine since that property is already in our dakota spec implicitly,
we don't allow aliasing of elements

Tested with all dk scenes on macos