thundr: Partial progress on dmabuf flickering
I think what's going on here is a culmination of various issues. One
of them just happens to be synchronizing the newly imported dmabuf,
which is done with an image barrier after importing that transfers
ownership to us from the foreign queue family.

I think the second problem, exposed by firefox's video player, is
something to do with damage and subsurfaces. Or subsurfaces and
popup windows. Something is definitely wrong with both of those.

edit: Originally I saw this remove some flickering, but it seems
to not have done anything. With a faster release build I can
still trigger it reliably. The webrender firefox setting doesn't
help, and there are some bugzilla reports about that having an
effect