thundr: Fix moltenvk
We needed to use the max_per_stage_descriptor_samplers, which defaults
to 16 with moltenvk as of writing, to properly reduce the number of
samplers that can be bound. This limit means that moltenvk is not
a suitable platform for thundr

It seems with the latest version the sampler count is bumped to 50000
or so, so that's an improvement. It may require binding samplers
and images separately?