thundr: make traditional composition the default
The compute pipeline is very power effective but cannot handle two
very common scenarios:
1. more than n (n = 4 by default) blendings
2. High window counts

This means it is really only useful for handling something like a
fullscreen surface with an overlay. As-is, this will stay disabled
until it makes sense to come back to it.

Tested with text, TT on nv/bsd