dakota: Remove format! usage in failure context calls
This is a funny gotcha: using format! in calls to anyhow's .context
call was causing the error string to *always* be formatted, not just
formatted when there is an error. Since I don't know how to prevent
this I just removed the formatting.

While here remove some excess prints from thundr that were getting
in the way.

Now resizing is waayyy smoother, flamegraph used to be 90% of time
spent formatting...