thundr: Begin to use viewports
This change adds a viewport argument so that we can control what part
of the screen we are drawing to at any point in time. This will be
used for dakota scrolling.

The viewports work in their own coordinate systems, just like the
normal full window view. The goal is to just submit surface lists
like normal and the viewport will take care of the rest.

Tested with DT on linux