ways: fix frame callback behavior
This fixes the frame callbacks, not sending only one per frame but
instead sending the entire list. Also marks the atmosphere as modified
after scroll events too, so that things are actually signaled.

Really this is temporary and we need to get rid of the dual thread
model. It's the biggest current mistake.