ways: append frame callbacks when committing
This also matches wlroots. Instead of clearing the currently committed
list of callbacks when new ones come in, we add the attached callbacks
to the committed list and let them all get signaled. This fixes most
of the firefox issues from the refactoring, although scrolling
stuttering and the occasional texture use after release still remain.