input: add wl_pointer fixes
Add logic while creating input events for wl_pointer to avoid querying
an axis that is not set. This also records the axis event source, to
send the axis_source event. Weirdly, the axis_source event does need
to be sent for firefox scrolling to work properly: it solves the can't
scroll all the way to the top/bottom of a page issue. With this change
however firefox stops performing scrolling when we visit a new webpage
after scrolling..