Rework thundr depth calculation and add renderdoc to dakota-test
Changing the axis depth is calculated on for thundr. It used to be
starting at 1.0 and decreasing for every object added, but now it
does the opposite. This makes it a little easier to understand, and
makes the code slightly cleaner.

Adding renderdoc to make it easier to record traces