dakota: Allow for different sized fonts
This adds a Font definition object, which lets the user define a
set of parameters to describe the font a text should load. This
can then be assigned to a text object to make all runs format
a certain way.

TODO: allow per-run font changes

Tested with DT, cat5 SDL on Linux