lluvia: merge Session into Component.
One of the problems with lluvia usage up the stack is that all sessions
are held inside dakota, meaning users are limited by borrow checking
rules on the main dakota reference instead of individual sessions. Since
we don't want to expose sessions to users, merge it into components and
we will expose those.