thundr: Remove the obsolete compute pipeline
This rendering method was interesting to pursue but does not handle
real world conditions. Mainly you need to guarantee no more than n
surfaces are layered or else transparency breaks. Also relied on
damage hints being honored which is technically not compliant.

Maybe some of these ideas can be recycled but for now this is getting
removed so it's not in the way of a thundr rework.