WIP: thundr+dakota: remove surfacelist
This removes the large surfaceslists that get populated and handed
to the drawing. They prevent incremental drawing and cause sync
headaches due to how they have to be updated.

This switches to have a draw call per element, with offscreen elements
not submitted for drawing. Includes some viewport rework to remove
the calculate_viewports pass.