lluvia: Finish removing locking from Snapshot
This stops using the parent Component constantly and instead gets
a read lock view of the internal table. The snapshot can then use
this table efficiently, and uses a writer lock when flushing.

This has a side effect of not allowing writes while a snapshot is
accurate. This is seen as a feature, where all updates are done
in a Snapshot without interacting with any locks, then flushed to
the Component on commit.