lluvia: Mark heavily used entry points as inline.
rustc does not do any cross-crate optimizations by default, which
has the nasty side effect of getting in our way with lluvia. We
try to open it up so that no locks are taken for Snapshots, but
due to the no cross-crate opt rule it never actually reduces a
Snapshot::get into a cheap memory access.

This change uses the inline tag to circumvent this. When functions
are marked with inline it provides the option of cross-crate optimization
to the compiler, which then does the right thing. This dramatically
helps debug performance and situations where LTO is not enabled.